Improve the verifier to allow to depends upon multiple version of the same NuGet
