+++ sonaranalyzer-dotnet/tests/SonarAnalyzer.UnitTest/NuGet.config
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<configuration>
+  <packageSources>
+    <add key="nuget.org" value="https://api.nuget.org/v3/index.json" />
+    <add key="nuget.org v2" value="https://www.nuget.org/api/v2/" />
+  </packageSources>
+</configuration>
