+++ sonaranalyzer-dotnet/tests/SonarAnalyzer.UnitTest/Rules/TestClassShouldHaveTestMethodTest.cs
@@ -27,24 +27,28 @@ namespace SonarAnalyzer.UnitTest.Rules
     [TestClass]
     public class TestClassShouldHaveTestMethodTest
     {
-        [TestMethod]
+        [DataTestMethod]
+        [DataRow("2.5.7.10213")]
+        [DataRow(AssemblyReference.NuGetInfo.LatestVersion)]
         [TestCategory("Rule")]
-        public void TestClassShouldHaveTestMethod_NUnit()
+        public void TestClassShouldHaveTestMethod_NUnit(string testFwkVersion)
         {
-            Verifier.VerifyAnalyzer(@"TestCases\TestClassShouldHaveTestMethod_NUnit.cs",
+            Verifier.VerifyAnalyzer(@"TestCases\TestClassShouldHaveTestMethod.NUnit.cs",
                 new TestClassShouldHaveTestMethod(),
                 null,
-                AssemblyReference.FromNuGet("nunit.framework.dll", "NUnit", "2.6.4"));
+                AssemblyReference.FromNuGet("nunit.framework.dll", "NUnit", testFwkVersion));
         }
 
-        [TestMethod]
+        [DataTestMethod]
+        [DataRow("1.1.11")]
+        [DataRow(AssemblyReference.NuGetInfo.LatestVersion)]
         [TestCategory("Rule")]
-        public void TestClassShouldHaveTestMethod_MSTest()
+        public void TestClassShouldHaveTestMethod_MSTest(string testFwkVersion)
         {
-            Verifier.VerifyAnalyzer(@"TestCases\TestClassShouldHaveTestMethod_MSTest.cs",
+            Verifier.VerifyAnalyzer(@"TestCases\TestClassShouldHaveTestMethod.MsTest.cs",
                 new TestClassShouldHaveTestMethod(),
                 null,
-                AssemblyReference.FromNuGet("Microsoft.VisualStudio.TestPlatform.TestFramework.dll", "MSTest.TestFramework", "1.2.0"));
+                AssemblyReference.FromNuGet("Microsoft.VisualStudio.TestPlatform.TestFramework.dll", "MSTest.TestFramework", testFwkVersion));
         }
     }
 }
