+++ sonaranalyzer-dotnet/tests/SonarAnalyzer.UnitTest/TestCases/AssertionArgsShouldBePassedInCorrectOrder.NUnit.cs
@@ -1,0 +1,29 @@
+﻿using System;
+using NUnit.Framework;
+
+namespace Tests.Diagnostics
+{
+    [TestFixture]
+    class Program
+    {
+        [Test]
+        public void Foo()
+        {
+            var str = "";
+            Assert.AreEqual(str, ""); // Noncompliant {{Make sure these 2 arguments are in the correct order: expected value, actual value.}}
+//                          ^^^^^^^
+            Assert.AreSame(str, ""); // Noncompliant
+//                         ^^^^^^^
+
+            double d = 42;
+            Assert.AreEqual(d, 42); // Noncompliant
+//                          ^^^^^
+            Assert.AreSame(d, 42); // Noncompliant
+            Assert.AreEqual(d, 42, 1, "message"); // Noncompliant
+
+            Assert.AreEqual("", str);
+            Assert.AreSame("", str);
+            Assert.AreEqual(42, d, 1, "message");
+        }
+    }
+}
