--- sonaranalyzer-dotnet/tests/SonarAnalyzer.UnitTest/TestCases/AssertionArgsShouldBePassedInCorrectOrder.cs
@@ -1,87 +1,0 @@
-﻿using System;
-
-namespace MsTestTests
-{
-    using Microsoft.VisualStudio.TestTools.UnitTesting;
-
-    [TestClass]
-    class Program
-    {
-        [TestMethod]
-        public void Foo()
-        {
-            var str = "";
-            Assert.AreEqual(str, ""); // Noncompliant {{Make sure these 2 arguments are in the correct order: expected value, actual value.}}
-//                          ^^^^^^^
-            Assert.AreSame(str, ""); // Noncompliant
-//                         ^^^^^^^
-
-            double d = 42;
-            Assert.AreEqual(d, 42); // Noncompliant
-//                          ^^^^^
-            Assert.AreSame(d, 42); // Noncompliant
-            Assert.AreEqual(d, 42, 1, "message"); // Noncompliant
-
-            Assert.AreEqual("", str);
-            Assert.AreSame("", str);
-            Assert.AreEqual(42, d, 1, "message");
-        }
-    }
-}
-
-namespace NUnitTests
-{
-    using NUnit.Framework;
-
-    [TestFixture]
-    class Program
-    {
-        [Test]
-        public void Foo()
-        {
-            var str = "";
-            Assert.AreEqual(str, ""); // Noncompliant {{Make sure these 2 arguments are in the correct order: expected value, actual value.}}
-//                          ^^^^^^^
-            Assert.AreSame(str, ""); // Noncompliant
-//                         ^^^^^^^
-
-            double d = 42;
-            Assert.AreEqual(d, 42); // Noncompliant
-//                          ^^^^^
-            Assert.AreSame(d, 42); // Noncompliant
-            Assert.AreEqual(d, 42, 1, "message"); // Noncompliant
-
-            Assert.AreEqual("", str);
-            Assert.AreSame("", str);
-            Assert.AreEqual(42, d, 1, "message");
-        }
-    }
-}
-
-namespace XUnitTests
-{
-    using Xunit;
-
-    [Fact]
-    class Program
-    {
-        public void Foo()
-        {
-            var str = "";
-            Assert.Equal(str, ""); // Noncompliant {{Make sure these 2 arguments are in the correct order: expected value, actual value.}}
-//                       ^^^^^^^
-            Assert.Same(str, ""); // Noncompliant
-//                      ^^^^^^^
-
-            double d = 42;
-            Assert.Equal(d, 42); // Noncompliant
-//                       ^^^^^
-            Assert.Same(d, 42); // Noncompliant
-            Assert.Equal(d, 42, 1, "message"); // Noncompliant
-
-            Assert.Equal("", str);
-            Assert.Same("", str);
-            Assert.Equal(42, d, 1, "message");
-        }
-    }
-}
