+++ sonaranalyzer-dotnet/tests/SonarAnalyzer.UnitTest/TestCases/ExpectedExceptionAttributeShouldNotBeUsed.MsTest.cs
@@ -1,4 +1,4 @@
-using System;
+﻿using System;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace Tests.Diagnostics
@@ -13,14 +13,6 @@ namespace Tests.Diagnostics
             x.ToString();
         }
 
-        [Test]
-        [NUnit.Framework.ExpectedException(typeof(ArgumentNullException))]  // Noncompliant
-        public void TestFoo2()
-        {
-            var x = true;
-            x.ToString();
-        }
-
         [TestMethod]
         [ExpectedException(typeof(ArgumentNullException))]  // Compliant - one line
         public void TestFoo3()
@@ -28,38 +20,12 @@ namespace Tests.Diagnostics
             new object().ToString();
         }
 
-        [Test]
-        [NUnit.Framework.ExpectedException(typeof(ArgumentNullException))]  // Compliant - one line
-        public void TestFoo4()
-        {
-            new object().ToString();
-        }
-
         [TestMethod]
         [ExpectedException(typeof(ArgumentNullException))]  // Compliant - one line
         public void TestFoo5 => new object().ToString();
-        }
-
-        [Test]
-        [NUnit.Framework.ExpectedException(typeof(ArgumentNullException))]  // Compliant - one line
-        public void TestFoo6 => new object().ToString();
 
         [TestMethod]
         public void TestFoo7()
-        {
-            bool callFailed = false;
-            try
-            {
-                //...
-            }
-            catch (ArgumentNullException)
-            {
-                callFailed = true;
-            }
-        }
-
-        [Test]
-        public void TestFoo8()
         {
             bool callFailed = false;
             try
