+++ sonaranalyzer-dotnet/tests/SonarAnalyzer.UnitTest/TestCases/TestMethodShouldHaveCorrectSignature.Xunit.cs
@@ -1,0 +1,34 @@
+﻿using System;
+using Xunit;
+
+namespace Tests.Diagnostics
+{
+    class XUnitTest
+    {
+
+        [Xunit.Fact]
+        private void PrivateTestMethod() // Compliant
+        {
+        }
+
+        [Fact]
+        protected void ProtectedTestMethod() // Compliant
+        {
+        }
+
+        [Fact]
+        internal void InternalTestMethod() // Compliant
+        {
+        }
+
+        [Fact]
+        public async void AsyncTestMethod()  // Noncompliant
+        {
+        }
+
+        [Fact]
+        public void GenericTestMethod<T>()  // Noncompliant
+        {
+        }
+    }
+}
